调整ga代码
