+++ ga.py
@@ -1,30 +1,16 @@
 # Google Analytics 上报脚本
-import platform
-import uuid
 from urllib.parse import quote_plus
 
 import requests
 
 from log import logger
-from util import try_except, get_screen_size
+from util import try_except, get_cid, get_resolution
 from version import now_version
-
-
-def get_cid():
-    return "{}-{}".format(platform.node(), uuid.getnode())
-
-
-def get_resolution() -> str:
-    width, height = get_screen_size()
-    return f"{width}x{height}"
-
-
-GA_REPORT_TYPE_EVENT = "event"
-GA_REPORT_TYPE_PAGE_VIEW = "page_view"
 
 # note: 查看数据地址 https://analytics.google.com/analytics/web/#/
 # note: 当发现上报失败时，可以将打印的post body复制到 https://ga-dev-tools.web.app/hit-builder/ 进行校验，看是否缺了参数，或者有参数不符合格式
 # note: 参数文档 https://developers.google.com/analytics/devguides/collection/protocol/v1/parameters
+GA_API_URL = "https://www.google-analytics.com/collect"
 GA_TRACKING_ID = "UA-179595405-1"
 
 headers = {
@@ -44,6 +30,9 @@ common_data = {
     'sr': get_resolution(),
 }
 
+GA_REPORT_TYPE_EVENT = "event"
+GA_REPORT_TYPE_PAGE_VIEW = "page_view"
+
 
 @try_except(show_exception_info=False)
 def track_event(category: str, action: str, label=None, value=0, ga_misc_params: dict = None):
@@ -62,7 +51,7 @@ def track_event(category: str, action: str, label=None, value=0, ga_misc_params:
         **ga_misc_params,  # 透传的一些额外参数
     }
 
-    res = requests.post('https://www.google-analytics.com/collect', data=data, headers=headers, timeout=10)
+    res = requests.post(GA_API_URL, data=data, headers=headers, timeout=10)
     logger.debug(f"request body = {res.request.body}")
 
 
@@ -83,7 +72,7 @@ def track_page(page: str, ga_misc_params: dict = None):
         **ga_misc_params,  # 透传的一些额外参数
     }
 
-    res = requests.post('https://www.google-analytics.com/collect', data=data, timeout=10)
+    res = requests.post(GA_API_URL, data=data, timeout=10)
     logger.debug(f"request body = {res.request.body}")
